Major Graphical Lift.
I was to tired to actually do this right yesterday, but I did lay the path for it, and today I 'finished' it. What I mean is that I could actually add in something to get it working better, with no changing the script by hand, but it's easy, and I want it to be automatic to the best of it's abilities.